modularize InputSource impl
